Clean code via sonarqube analysis

+++ Indy3DModInstaller/Indy3DModInstaller.cs
@@ -10,29 +10,41 @@ internal class Indy3DRegistryEntry(string gameVersionId, string registryKey)
     public string RegistryKey { get; set; } = registryKey;
 }
 
-internal class Indy3DModInstaller
+internal class Indy3DModInstaller(IMessageWriter messageWriter)
 {
-    private static readonly Indy3DRegistryEntry[] _registryEntries = [
+    private static readonly Indy3DRegistryEntry[] REGISTRY_ENTRIES = [
         new Indy3DRegistryEntry("Steam", "HKEY_LOCAL_MACHINE\\SOFTWARE\\WOW6432Node\\LucasArts Entertainment Company LLC\\Indiana Jones and the Infernal Machine\\v1.0"),
         new Indy3DRegistryEntry("GOG", "HKEY_CURRENT_USER\\SOFTWARE\\LucasArts Entertainment Company LLC\\Indiana Jones and the Infernal Machine\\v1.0"),
         new Indy3DRegistryEntry("CD", "HKEY_LOCAL_MACHINE\\Software\\LucasArts Entertainment Company LLC\\Indiana Jones and the Infernal Machine\\v1.0")
     ];
 
-    private static readonly string[] _gameAssetFolderNames = ["3do", "cog", "hi3do", "mat", "misc", "ndy", "sound"];
+    private static readonly string[] GAME_ASSET_FOLDER_NAMES = ["3do", "cog", "hi3do", "mat", "misc", "ndy", "sound"];
 
-    private const string _cogBackupFolder = "cog_backup";
+    private const string RESOURCE_FOLDER = "Resource";
+    private const string COG_BACKUP_FOLDER = "cog_backup";
 
-    private const string _jones3dGobFile = "Jones3D.GOB";
-    private const string _jones3dGobBackupFile = "Jones3D.GOB.BAK";
-    private const string _cd1GobFile = "CD1.GOB";
-    private const string _cd1GobBackupFile = "CD1.GOB.BAK";
-    private const string _cd2GobFile = "CD2.GOB";
-    private const string _cd2GobBackupFile = "CD2.GOB.BAK";
+    private const string JONES3D_GOB_FILE = "Jones3D.GOB";
+    private const string JONES3D_GOB_BACKUP_FILE = $"{JONES3D_GOB_FILE}.BAK";
+    private const string CD1_GOB_FILE = "CD1.GOB";
+    private const string CD1_GOB_BACKUP_FILE = $"{CD1_GOB_FILE}.BAK";
+    private const string CD2_GOB_FILE = "CD2.GOB";
+    private const string CD2_GOB_BACKUP_FILE = $"{CD2_GOB_FILE}.BAK";
 
-    public static void Unpack(string installPath)
+    private readonly IMessageWriter _messageWriter = messageWriter;
+
+    public void Unpack(string? installPath)
     {
+        if (installPath == null)
+        {
+            throw new ArgumentNullException($"ERROR: Path empty. Cannot unpack game files.{Environment.NewLine}Please select path to Resource folder.");
+        }
+        else if (Path.GetFileName(installPath) != RESOURCE_FOLDER)
+        {
+            throw new ArgumentException($"ERROR: Path doesn't lead to Resource folder. Cannot unpack game files.{Environment.NewLine}Please select path to Resource folder.");
+        }
+
         string cogPath = Path.Combine(installPath, "cog");
-        string cogBackupPath = Path.Combine(installPath, _cogBackupFolder);
+        string cogBackupPath = Path.Combine(installPath, COG_BACKUP_FOLDER);
 
         // Backup existing cog override (important for Steam version)
         if (Directory.Exists(cogPath))
@@ -40,27 +52,20 @@ internal class Indy3DModInstaller
             Directory.Move(cogPath, cogBackupPath);
         }
 
-        try
-        {
-            Program.WriteLine($"Extracting archive {_jones3dGobFile}...");
-            OsUtils.LaunchProcess("gobext.exe", [_jones3dGobFile, "-o=."], installPath);
-            Program.WriteLine($"Extracting archive {_cd1GobFile}...");
-            OsUtils.LaunchProcess("gobext.exe", [_cd1GobFile, "-o=."], installPath);
-            Program.WriteLine($"Extracting archive {_cd2GobFile}...");
-            OsUtils.LaunchProcess("gobext.exe", [_cd2GobFile, "-o=."], installPath);
-        }
-        catch (Exception)
-        {
-            throw;
-        }
+        _messageWriter.WriteLine($"Extracting archive {JONES3D_GOB_FILE}...");
+        OsUtils.LaunchProcess("gobext.exe", [JONES3D_GOB_FILE, "-o=."], installPath);
+        _messageWriter.WriteLine($"Extracting archive {CD1_GOB_FILE}...");
+        OsUtils.LaunchProcess("gobext.exe", [CD1_GOB_FILE, "-o=."], installPath);
+        _messageWriter.WriteLine($"Extracting archive {CD2_GOB_FILE}...");
+        OsUtils.LaunchProcess("gobext.exe", [CD2_GOB_FILE, "-o=."], installPath);
 
-        string jones3dPath = Path.Combine(installPath, _jones3dGobFile);
-        string cd1Path = Path.Combine(installPath, _cd1GobFile);
-        string cd2Path = Path.Combine(installPath, _cd2GobFile);
+        string jones3dPath = Path.Combine(installPath, JONES3D_GOB_FILE);
+        string cd1Path = Path.Combine(installPath, CD1_GOB_FILE);
+        string cd2Path = Path.Combine(installPath, CD2_GOB_FILE);
 
-        string jones3dBackupPath = Path.Combine(installPath, _jones3dGobBackupFile);
-        string cd1BackupPath = Path.Combine(installPath, _cd1GobBackupFile);
-        string cd2BackupPath = Path.Combine(installPath, _cd2GobBackupFile);
+        string jones3dBackupPath = Path.Combine(installPath, JONES3D_GOB_BACKUP_FILE);
+        string cd1BackupPath = Path.Combine(installPath, CD1_GOB_BACKUP_FILE);
+        string cd2BackupPath = Path.Combine(installPath, CD2_GOB_BACKUP_FILE);
 
         // rename/backup GOB files so that the game is running solely from extracted files
         File.Move(jones3dPath, jones3dBackupPath);
@@ -68,17 +73,11 @@ internal class Indy3DModInstaller
         File.Move(cd2Path, cd2BackupPath);
 
         string[] cnd_files = Directory.GetFiles(Path.Combine(installPath, "ndy"), "*.cnd");
-        try
+        
+        foreach (string file in cnd_files)
         {
-            foreach (string file in cnd_files)
-            {
-                Program.WriteLine($"Extracting level {Path.GetFileName(file)}...");
-                OsUtils.LaunchProcess("cndtool.exe", ["extract", "--no-template", $"-o=.", $"{Path.Combine("ndy", Path.GetFileName(file))}"], installPath);
-            }
-        }
-        catch (Exception)
-        {
-            throw;
+            _messageWriter.WriteLine($"Extracting level {Path.GetFileName(file)}...");
+            OsUtils.LaunchProcess("cndtool.exe", ["extract", "--no-template", $"-o=.", $"{Path.Combine("ndy", Path.GetFileName(file))}"], installPath);
         }
 
         string keyPath = Path.Combine(installPath, "key");
@@ -95,11 +94,13 @@ internal class Indy3DModInstaller
         {
             OsUtils.CopyDirectoryContent(cogBackupPath, cogPath);
         }
+
+        _messageWriter.WriteLine("Unpacking successfully finished.");
     }
 
     public static string? GetInstallPathFromRegistry()
     {
-        foreach (Indy3DRegistryEntry registryEntry in _registryEntries)
+        foreach (Indy3DRegistryEntry registryEntry in REGISTRY_ENTRIES)
         {
             object? registryKey = Registry.GetValue(registryEntry.RegistryKey, "Install Path", null);
             if (registryKey == null)
@@ -114,9 +115,9 @@ internal class Indy3DModInstaller
         return null;
     }
 
-    public static void SetDevMode()
+    public void SetDevMode()
     {
-        foreach (Indy3DRegistryEntry registryEntry in _registryEntries)
+        foreach (Indy3DRegistryEntry registryEntry in REGISTRY_ENTRIES)
         {
             object? registryKey = Registry.GetValue(registryEntry.RegistryKey, "Start Mode", 42);
             if (registryKey == null)
@@ -124,27 +125,41 @@ internal class Indy3DModInstaller
                 continue;
             }
 
-            Program.WriteLine($"Dev Mode: Found entry for {registryEntry.GameVersionId} version.");
+            _messageWriter.WriteLine($"Dev Mode: Found entry for {registryEntry.GameVersionId} version.");
             int startMode = (int) registryKey;
             if (startMode != 2)
             {
                 Registry.SetValue(registryEntry.RegistryKey, "Start Mode", 2, RegistryValueKind.DWord);
-                Program.WriteLine("Dev Mode for Indy3D.exe enabled.");
+                _messageWriter.WriteLine("Dev Mode for Indy3D.exe enabled.");
             }
             else
             {
                 Registry.SetValue(registryEntry.RegistryKey, "Start Mode", 1, RegistryValueKind.DWord);
-                Program.WriteLine("Dev Mode was already enabled.");
-                Program.WriteLine("Dev Mode for Indy3D.exe disabled.");
+                _messageWriter.WriteLine("Dev Mode was already enabled.");
+                _messageWriter.WriteLine("Dev Mode for Indy3D.exe disabled.");
             }
         }
     }
 
-    public static void Install(string installPath, string modPath)
+    public void Install(string? installPath, string? modPath)
     {
-        Program.WriteLine($"Installing mod from {modPath}...");
+        if (installPath == null)
+        {
+            throw new ArgumentNullException($"ERROR: Path empty. Cannot install mod.{Environment.NewLine}Please select path to Resource folder.");
+        }
+        else if (Path.GetFileName(installPath) != RESOURCE_FOLDER)
+        {
+            throw new ArgumentException($"ERROR: Path doesn't lead to Resource folder. Cannot install mod.{Environment.NewLine}Please select path to Resource folder.");
+        }
 
-        foreach (string folderName in _gameAssetFolderNames)
+        if (modPath == null)
+        {
+            throw new ArgumentNullException($"ERROR: Path empty. Cannot install mod.{Environment.NewLine}Please select path to mod folder.");
+        }
+
+        _messageWriter.WriteLine($"Installing mod from {modPath}...");
+
+        foreach (string folderName in GAME_ASSET_FOLDER_NAMES)
         {
             string folderInInstallPath = Path.Combine(installPath, folderName);
             string folderInModPath = Path.Combine(modPath, folderName);
@@ -152,34 +167,45 @@ internal class Indy3DModInstaller
             if (Directory.Exists(folderInModPath))
             {
                 string folderinModPathWithModPrefix = Path.Combine(Path.GetFileName(modPath), Path.GetFileName(folderInModPath));
-                Program.WriteLine($"Installing {folderinModPathWithModPrefix}...");
+                _messageWriter.WriteLine($"Installing {folderinModPathWithModPrefix}...");
                 OsUtils.CopyDirectoryContent(folderInModPath, folderInInstallPath);
             }
         }
+
+        _messageWriter.WriteLine("Mod installation successfully finished.");
     }
 
-    public static void Uninstall(string installPath)
+    public void Uninstall(string? installPath)
     {
-        Program.WriteLine("Uninstalling mods, reverting to vanilla state from backups...");
+        if (installPath == null)
+        {
+            throw new ArgumentNullException($"ERROR: Path empty. Cannot uninstall mods.{Environment.NewLine}Please select path to Resource folder.");
+        }
+        else if (Path.GetFileName(installPath) != RESOURCE_FOLDER)
+        {
+            throw new ArgumentException($"ERROR: Path doesn't lead to Resource folder. Cannot uninstall mods.{Environment.NewLine}Please select path to Resource folder.");
+        }
 
-        foreach (string folderName in _gameAssetFolderNames)
+        _messageWriter.WriteLine("Uninstalling mods, reverting to vanilla state from backups...");
+
+        foreach (string folderName in GAME_ASSET_FOLDER_NAMES)
         {
             string folderInInstallPath = Path.Combine(installPath, folderName);
 
             if (Directory.Exists(folderInInstallPath))
             {
-                Program.WriteLine($"Uninstalling modded {Path.GetFileName(folderInInstallPath)}...");
+                _messageWriter.WriteLine($"Uninstalling modded {Path.GetFileName(folderInInstallPath)}...");
                 Directory.Delete(folderInInstallPath, true);
             }
         }
 
-        string jones3dPath = Path.Combine(installPath, _jones3dGobFile);
-        string cd1Path = Path.Combine(installPath, _cd1GobFile);
-        string cd2Path = Path.Combine(installPath, _cd2GobFile);
+        string jones3dPath = Path.Combine(installPath, JONES3D_GOB_FILE);
+        string cd1Path = Path.Combine(installPath, CD1_GOB_FILE);
+        string cd2Path = Path.Combine(installPath, CD2_GOB_FILE);
 
-        string jones3dBackupPath = Path.Combine(installPath, _jones3dGobBackupFile);
-        string cd1BackupPath = Path.Combine(installPath, _cd1GobBackupFile);
-        string cd2BackupPath = Path.Combine(installPath, _cd2GobBackupFile);
+        string jones3dBackupPath = Path.Combine(installPath, JONES3D_GOB_BACKUP_FILE);
+        string cd1BackupPath = Path.Combine(installPath, CD1_GOB_BACKUP_FILE);
+        string cd2BackupPath = Path.Combine(installPath, CD2_GOB_BACKUP_FILE);
 
         // Restore backups
         if (File.Exists(jones3dBackupPath))
@@ -198,18 +224,31 @@ internal class Indy3DModInstaller
         }
 
         string cogPath = Path.Combine(installPath, "cog");
-        string cogBackupPath = Path.Combine(installPath, _cogBackupFolder);
+        string cogBackupPath = Path.Combine(installPath, COG_BACKUP_FOLDER);
 
         if (Directory.Exists(cogBackupPath))
         {
             Directory.Move(cogBackupPath, cogPath);
         }
+
+        _messageWriter.WriteLine("Mod uninstallation successfully finished.");
     }
 
-    public static void LaunchGame(string installPath)
+    public void LaunchGame(string? installPath)
     {
-        Program.WriteLine("Launching game...");
+        if (installPath == null)
+        {
+            throw new ArgumentNullException($"ERROR: Path empty. Cannot launch game.{Environment.NewLine}Please select path to Resource folder.");
+        }
+        else if (Path.GetFileName(installPath) != "Resource")
+        {
+            throw new ArgumentException($"ERROR: Path doesn't lead to Resource folder. Cannot launch game.{Environment.NewLine}Please select path to Resource folder.");
+        }
+
+        _messageWriter.WriteLine("Launching game...");
 
         OsUtils.LaunchProcess(Path.Combine(installPath, "Indy3D.exe"), [], installPath);
+
+        _messageWriter.WriteLine("Game exited successfully.");
     }
 }
